refactor: convert cart page to server component
- Convert CartPage from client to server component using getTranslations
- Move authentication logic from page to CartItemsList component
- Remove userId prop from CartItemsList, use auth hook internally
- Hide CheckOut component when cart is empty
- Simplify cart page structure by removing conditional rendering logic
